students/Khisametdinova_Dinara/Lab_3/parser/requirements.txt
fastapi
celery
redis
celery[beat]
aiohttp
beautifulsoup4
psycopg2-binary
uvicorn
python-dotenv==1.0.1